+++ src/main/java/org/truffleruby/core/module/ModuleFields.java
@@ -20,6 +20,7 @@ import com.oracle.truffle.api.utilities.CyclicAssumption;
 
 import org.truffleruby.Layouts;
 import org.truffleruby.RubyLanguage;
+import org.truffleruby.collections.ConcurrentOperations;
 import org.truffleruby.RubyContext;
 import org.truffleruby.core.klass.ClassNodes;
 import org.truffleruby.core.method.MethodFilter;
@@ -320,12 +321,13 @@ public class ModuleFields extends ModuleChain implements ObjectGraphNode {
     @TruffleBoundary
     public void setAutoloadConstant(RubyContext context, Node currentNode, String name, DynamicObject filename) {
         assert RubyGuards.isRubyString(filename);
+        RubyConstant autoloadConstant = setConstantInternal(context, currentNode, name, filename, true);
         if (context.getOptions().LOG_AUTOLOAD) {
-            RubyLanguage.LOGGER.info(() -> String.format("%s: setting up autoload %s::%s with %s",
+            RubyLanguage.LOGGER.info(() -> String.format("%s: setting up autoload %s with %s",
                     context.fileLine(context.getCallStack().getTopMostUserSourceSection()),
-                    getName(), name, filename));
+                    autoloadConstant, filename));
         }
-        setConstantInternal(context, currentNode, name, filename, true);
+        context.getFeatureLoader().addAutoload(autoloadConstant);
     }
 
     private RubyConstant setConstantInternal(RubyContext context, Node currentNode, String name, Object value, boolean autoload) {
@@ -333,22 +335,22 @@ public class ModuleFields extends ModuleChain implements ObjectGraphNode {
 
         SharedObjects.propagate(context, rubyModuleObject, value);
 
-        while (true) {
-            final RubyConstant previous = constants.get(name);
-            final boolean isPrivate = previous != null && previous.isPrivate();
-            final boolean isDeprecated = previous != null && previous.isDeprecated();
-            final SourceSection sourceSection = currentNode != null ?  currentNode.getSourceSection() : null;
-            final RubyConstant newValue = new RubyConstant(rubyModuleObject, value, isPrivate, autoload, isDeprecated, sourceSection);
+        RubyConstant previous;
+        RubyConstant newConstant;
+        do {
+            previous = constants.get(name);
+            newConstant = newConstant(currentNode, name, value, autoload, previous);
+        } while (!ConcurrentOperations.replace(constants, name, previous, newConstant));
 
-            if ((previous == null) ? (constants.putIfAbsent(name, newValue) == null) : constants.replace(name, previous, newValue)) {
-                newConstantsVersion();
-                if (!autoload && previous != null && !previous.isAutoload()) {
-                    return previous;
-                } else {
-                    return null;
-                }
-            }
-        }
+        newConstantsVersion();
+        return autoload ? newConstant : previous;
+    }
+
+    private RubyConstant newConstant(Node currentNode, String name, Object value, boolean autoload, RubyConstant previous) {
+        final boolean isPrivate = previous != null && previous.isPrivate();
+        final boolean isDeprecated = previous != null && previous.isDeprecated();
+        final SourceSection sourceSection = currentNode != null ?  currentNode.getSourceSection() : null;
+        return new RubyConstant(rubyModuleObject, name, value, isPrivate, autoload, isDeprecated, sourceSection);
     }
 
     @TruffleBoundary