+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Changes:
   unsupported, have been removed.
 * Our experimental JRuby-compatible Java interop has been removed - use
   `Polyglot` and `Java` instead.
+Bug fixes:
+
+* `autoload :C, "path"; require "path"` now correctly triggers the autoload.
 
 # 1.0 RC 14